Eliminated a "signed vs. unsigned comparison" compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@4287 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/blocks/states.cpp
@@ -142,7 +142,7 @@ bool States::DecodeStates ()
 {
   int i, numSections;
   // Find the floats.
-  for(i = 0, numSections = 0; i < ST_ENCODED_LENGTH; i++)
+  for(i = 0, numSections = 0; i < (int)ST_ENCODED_LENGTH; i++)
   {
     if (encodedData[i] == ',')
     {
